+++ model/src/main/java/agency/amazon/test/model/TrafficByDate.java
@@ -2,6 +2,8 @@ package agency.amazon.test.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TrafficByDate {
     private int browserPageViews;
@@ -16,12 +18,12 @@ public class TrafficByDate {
     private int mobileAppSessionsB2B;
     private int sessions;
     private int sessionsB2B;
-    private int buyBoxPercentage;
-    private int buyBoxPercentageB2B;
-    private int orderItemSessionPercentage;
-    private int orderItemSessionPercentageB2B;
-    private int unitSessionPercentage;
-    private int unitSessionPercentageB2B;
+    private BigDecimal buyBoxPercentage;
+    private BigDecimal buyBoxPercentageB2B;
+    private BigDecimal orderItemSessionPercentage;
+    private BigDecimal orderItemSessionPercentageB2B;
+    private BigDecimal unitSessionPercentage;
+    private BigDecimal unitSessionPercentageB2B;
     private int averageOfferCount;
     private int averageParentItems;
     private int feedbackReceived;
